backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
SQLAlchemy>=2.0.0
asyncpg>=0.27.0
alembic>=1.10.0
pydantic>=2.0
python-dotenv>=1.0
celery[redis]>=5.2.0
redis>=4.5.0
boto3>=1.26.0
aiohttp>=3.8.0
httpx>=0.24.0
python-multipart>=0.0.6
pytest>=7.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
assemblyai>=0.19.0
